move buffers into codegen for eventual CBufferFormat type
