--- src/finch/buffer/__init__.py
@@ -1,6 +1,0 @@
-from .buffer import AbstractBuffer, NumpyBuffer
-
-__all__ = [
-    "AbstractBuffer",
-    "NumpyBuffer",
-]
