+++ src/finch/codegen/__init__.py
@@ -1,3 +1,8 @@
-from .c import CKernel, get_c_function
+from .c import AbstractBuffer, CKernel, NumpyBuffer, get_c_function
 
-__all__ = ["get_c_function", "CKernel"]
+__all__ = [
+    "get_c_function",
+    "CKernel",
+    "AbstractBuffer",
+    "NumpyBuffer",
+]
